- bugfix for conditional logic error when javascript is disabled
